// src/types.tsx
export type CityData = {
  city: string;
  city_ascii: string;
  state_id: string;
  state_name: string;
  county_fips: string;
  county_name: string;
  lat: string;
  lng: string;
  population: string;
  density: string;
  source: string;
  military: string;
  incorporated: string;
  timezone: string;
  ranking: string;
  zips: string;
  id: string;
};